fixed time_t and size_t abuse

+++ GrawToROOT/src/EventSourceGRAW.cpp
@@ -336,7 +336,7 @@ void EventSourceGRAW::fillEventRawFromFrame(GET::GDataFrame & aGrawFrame){
   // no pedestal subtracion (can be done later using raw data stored in the container class
 
   myCurrentEventRaw->SetEventId((uint64_t)aGrawFrame.fHeader.fEventIdx);
-  myCurrentEventRaw->SetEventTimestamp((uint32_t)aGrawFrame.fHeader.fEventTime);
+  myCurrentEventRaw->SetEventTimestamp(aGrawFrame.fHeader.fEventTime);
   
   uint8_t COBO_idx = (uint8_t)aGrawFrame.fHeader.fCoboIdx;
   uint8_t ASAD_idx = (uint8_t)aGrawFrame.fHeader.fAsadIdx;
